refactor(ALL-1461): convert to custom hook
